mas alertas y detalles de front

+++ ABM_Entradas.php
@@ -63,7 +63,7 @@ include_once('db/conexionDB.php');
             <td> <?php echo $fila['cantidad'];?></td>
             <td> <?php echo $fila2['fechaEvento'];?></td>
             <td><a href="actualizar_entrada.php?id=<?php echo $fila['id'] ?>" class="btn btn-info">Editar</a></td>
-            <td><a href="delete_entrada.php?id=<?php echo $fila['id'] ?>" class="btn btn-danger">Eliminar</a></td>              
+            <td><a href="delete_entrada.php?id=<?php echo $fila['id'] ?>" class="btn btn-danger" onclick="return ConfirmDelete()"> Eliminar  </a></td>              
           </tr>
         <?php
             } }; 
@@ -71,5 +71,28 @@ include_once('db/conexionDB.php');
         </table>
           </div>
 </div>
+
+
+<script type="text/javascript">
+        function ConfirmDelete(){
+          var respuesta = confirm('Esta seguro que desea eliminar?');
+       if (respuesta== true) {
+          return true;
+        }else {
+          return false;
+        }
+      }
+    </script>
+<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
+<script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js" integrity="sha384-9/reFTGAW83EW2RDu2S0VKaIzap3H66lZH81PoYlFhbGU+6BZp6G7niu735Sk7lN" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/js/bootstrap.min.js" integrity="sha384-w1Q4orYjBQndcko6MimVbzY0tgp4pWB4lZ7lr30WKz0vr/aWKhXdBNmNb5D92v7s" crossorigin="anonymous"></script>
+
+
+
+
+
+
+
 </body>
 </html>
