+++ listadoEventos.php
@@ -48,7 +48,7 @@ while ($fila2 = mysqli_fetch_array($resultado2)) {
   }
      ?> </td>
      <td>
-<a class="trash"  class="btn btn-outline-danger" href="eliminarEvento.php?id=<?php echo $idPublicacion; ?>"><i class='fa-regular fa-trash-can'></i></a>
+     <a class="trash"  class="btn btn-outline-danger" href="eliminarEvento.php?id=<?php echo $idPublicacion; ?>"><i class='fa-regular fa-trash-can' onclick="return ConfirmDelete()"></i></a>
 <a class='pencil' class='btn btn-outline-success' href="eventoABM2.php?id=<?php echo $idPublicacion; ?>"><i class='fa-solid fa-pencil'></i></a> 
 <?php 
 
@@ -65,6 +65,16 @@ if (existe_entrada($idPublicacion, $conexion)) {
   </tbody>
 </table>
 </div>
+<script type="text/javascript">
+        function ConfirmDelete(){
+          var respuesta = confirm('Esta seguro que desea eliminar?');
+       if (respuesta== true) {
+          return true;
+        }else {
+          return false;
+        }
+      }
+    </script>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
 <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js" integrity="sha384-9/reFTGAW83EW2RDu2S0VKaIzap3H66lZH81PoYlFhbGU+6BZp6G7niu735Sk7lN" crossorigin="anonymous"></script>
