+++ listadoPublicaciones.php
@@ -50,7 +50,7 @@ while ($fila2 = mysqli_fetch_array($resultado2)) {
      ?> </td>
 <td>
 <a class='pencil'  href='publicacionABM.php?id=<?php echo $fila['idPublicacion'] ?>'><i class='fa-solid fa-pencil'></i></a>
-<a  class='trash' href='eliminarPublicacion.php?id=<?php echo $fila['idPublicacion'] ?>'><i class='fa-regular fa-trash-can'></i></a>
+<a  class='trash' href='eliminarPublicacion.php?id=<?php echo $fila['idPublicacion'] ?>'><i class='fa-regular fa-trash-can' onclick="return ConfirmDelete()"></i><a>
 </td>
 
 </td>
@@ -60,7 +60,18 @@ while ($fila2 = mysqli_fetch_array($resultado2)) {
   </tbody>
 </table>
 </div>
+<script type="text/javascript">
+        function ConfirmDelete(){
+          var respuesta = confirm('Esta seguro que desea eliminar?');
+       if (respuesta== true) {
+          return true;
+        }else {
+          return false;
+        }
+      }
+    </script>
 <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.1.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-ka7Sk0Gln4gmtz2MlQnikT1wXgYsOg+OMhuP+IlRH9sENBO0LRn5q+8nbTov4+1p" crossorigin="anonymous"></script>
 <script src="https://code.jquery.com/jquery-3.5.1.slim.min.js" integrity="sha384-DfXdz2htPH0lsSSs5nCTpuj/zy4C+OGpamoFVy38MVBnE+IbbVYUew+OrCXaRkfj" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/popper.js@1.16.1/dist/umd/popper.min.js" integrity="sha384-9/reFTGAW83EW2RDu2S0VKaIzap3H66lZH81PoYlFhbGU+6BZp6G7niu735Sk7lN" crossorigin="anonymous"></script>
     <script src="https://cdn.jsdelivr.net/npm/bootstrap@4.5.3/dist/js/bootstrap.min.js" integrity="sha384-w1Q4orYjBQndcko6MimVbzY0tgp4pWB4lZ7lr30WKz0vr/aWKhXdBNmNb5D92v7s" crossorigin="anonymous"></script>
+
